Changed EpicSeeder to SpaceSeeder and remvoed Logs from cached git, as well as adding Logs to gitignore
