+++ .gitignore
@@ -351,6 +351,9 @@ _ReSharper*/
 
 *.jfm
 
+# Log Files
+/PersonalHub.Api/Logs
+
 #======================================================================================
 # The below section could possibly be removed as these should be ignored by the above.
 #======================================================================================
