--- PersonalHub.Api/Logs/log-20240820.txt
@@ -1,6 +1,0 @@
-2024-08-20 21:30:48.310 +02:00 [WRN] The ASP.NET Core developer certificate is not trusted. For information about trusting the ASP.NET Core developer certificate, see https://aka.ms/aspnet/https-trust-dev-cert.
-2024-08-20 21:30:48.428 +02:00 [INF] Now listening on: https://localhost:7149
-2024-08-20 21:30:48.429 +02:00 [INF] Now listening on: http://localhost:5111
-2024-08-20 21:30:48.430 +02:00 [INF] Application started. Press Ctrl+C to shut down.
-2024-08-20 21:30:48.430 +02:00 [INF] Hosting environment: Development
-2024-08-20 21:30:48.430 +02:00 [INF] Content root path: C:\Users\zippe\source\repos\PersonalHub\PersonalHub.Api
